+++ src/platform/mod.rs
@@ -53,10 +53,12 @@ pub mod user;
 
 pub mod ansi;
 pub mod command;
+pub mod download;
 pub mod environment;
 pub mod ipc;
 pub mod notifications;
 pub mod process;
+pub mod progress;
 pub mod server_lifecycle;
 pub mod shell_integration;
 pub mod sound;